+++ Tests/Integration/CheckoutRequest/testcases/express/caseDoExpressCheckoutPayment_Automatic_Brutto.php
@@ -130,6 +130,7 @@ $data = array(
         'sOEPayPalTransactionMode'  => 'Automatic',
         'sOEPayPalEmptyStockLevel'  => 1,
     ),
+    'session' => ['oepaypal' => 2],
     'expected'  => array(
         'requestToPayPal' => array(
             'VERSION'                            => '84.0',
@@ -144,7 +145,7 @@ $data = array(
             'PAYMENTREQUEST_0_NOTIFYURL'         => '{SHOP_URL}index.php?cl=oepaypalipnhandler&fnc=handleRequest&shp={SHOP_ID}',
             'PAYMENTREQUEST_0_DESC'              => 'Bestellnummer 1',
             'PAYMENTREQUEST_0_CUSTOM'            => 'Bestellnummer 1',
-            'BUTTONSOURCE'                       => '{BN_ID}',
+            'BUTTONSOURCE'                       => '{BN_ID_SHORTCUT}',
             'METHOD'                             => 'DoExpressCheckoutPayment',
             'PAYMENTREQUEST_0_SHIPTONAME'        => 'John Doe',
             'PAYMENTREQUEST_0_SHIPTOSTREET'      => 'Maple Street 10',