+++ Tests/Integration/CheckoutRequest/testcases/standard/caseDoExpressCheckoutPayment_Authorization_Netto.php
@@ -76,6 +76,7 @@ $data = array(
         'sOEPayPalTransactionMode'  => 'Authorization',
         'blShowNetPrice'            => true,
     ),
+    'session' => ['oepaypal' => 1],
     'expected'  => array(
         'requestToPayPal' => array(
             'VERSION'                            => '84.0',