+++ Tests/Integration/CheckoutRequest/testcases/standard/caseDoExpressCheckoutPayment_UserLoggedWithAddress.php
@@ -94,6 +94,7 @@ $data = array(
     ),
     'session'   => array(
         'deladrid' => 'TestUserAddressId',
+        'oepaypal' => 1,
     ),
     'expected'  => array(
         'requestToPayPal' => array(